break up statistics.py into pieces
