+++ datacube_stats/statistics/uncategorized.py
@@ -1,111 +1,25 @@
-"""
-Classes for performing statistical data analysis.
-"""
-import logging
-import abc
-import sys
-import os
+import warnings
 
 from collections import OrderedDict, Sequence
 from functools import partial
 from datetime import datetime
-from typing import Iterable
-from pydoc import locate
-from copy import copy
 
 import numpy as np
 import xarray
 
-from pkg_resources import iter_entry_points
 from datacube.model import Measurement
-from datacube.storage.masking import create_mask_value
 from datacube_stats.utils.dates import datetime64_to_inttime
-from .incremental_stats import (mk_incremental_sum, mk_incremental_or,
-                                compose_proc, broadcast_proc)
-from .utils import da_nodata, mk_masker, first_var
-from .stat_funcs import axisindex, argpercentile, _compute_medoid
-from .stat_funcs import anynan, section_by_index, medoid_indices
+from datacube_stats.utils import da_nodata
+from datacube_stats.stat_funcs import axisindex, argpercentile, _compute_medoid
+from datacube_stats.stat_funcs import anynan, section_by_index, medoid_indices
 
-
-LOG = logging.getLogger(__name__)
-
-
-class StatsConfigurationError(RuntimeError):
-    pass
-
-
-class StatsProcessingError(RuntimeError):
-    pass
-
-
-class Statistic(abc.ABC):
-    # pylint: disable=useless-return
-    @abc.abstractmethod
-    def compute(self, data: xarray.Dataset) -> xarray.Dataset:
-        """
-        Compute a statistic on the given Dataset.
-
-        # FIXME: Explain a little bit better, Dataset in, Dataset out, measurements match measurements()
-
-        :param xarray.Dataset data:
-        :return: xarray.Dataset
-        """
-
-    def measurements(self, input_measurements: Iterable[Measurement]) -> Iterable[Measurement]:
-        """
-        Turn a list of input measurements into a list of output measurements.
-
-        Base implementation simply copies input measurements to output_measurements.
-
-        # FIXME: Explain the purpose of this
-
-        :rtype: List[Measurement]
-        """
-        return input_measurements
-
-    def is_iterative(self) -> bool:
-        """
-        Should return True if class supports iterative computation one time slice at a time.
-
-        :rtype: Bool
-        """
-        return False
-
-    def make_iterative_proc(self):
-        """
-        Should return `None` if `is_iterative()` returns `False`.
-
-        Should return processing function `proc` that closes over internal
-        state that get updated one time slice at time, if `is_iterative()`
-        returns `True`.
-
-        proc(dataset_slice)  # Update internal state, called many times
-        result = proc()  # Extract final result, called once
-
-
-        See `incremental_stats.assemble_updater`
-
-        """
+from .core import Statistic, PerPixelMetadata, SimpleStatistic
+from .core import StatsProcessingError, StatsConfigurationError
 
 
 class NoneStat(Statistic):
     def compute(self, data):
         return data
-
-
-class SimpleStatistic(Statistic):
-    """
-    Describes the outputs of a statistic and how to calculate it
-
-    :param stat_func:
-        callable to compute statistics. Should both accept and return a :class:`xarray.Dataset`.
-    """
-
-    def __init__(self, stat_func):
-        self.stat_func = stat_func
-
-    def compute(self, data):
-        return self.stat_func(data)
 
 
 class ReducingXarrayStatistic(Statistic):
@@ -469,24 +383,6 @@ class Percentile(PerBandIndexStat):
         return PerBandIndexStat(per_pixel_metadata=self.per_pixel_metadata).measurements(renamed)
 
 
-class PerPixelMetadata(abc.ABC):
-    # pylint: disable=useless-return
-    def __init__(self, var_name='observed'):
-        self._var_name = var_name
-
-    @abc.abstractmethod
-    def compute(self, data, selected_indexes):
-        """
-        Return a variable name and :class:`xarray.Variable` to add in to the dataset.
-        """
-
-    @abc.abstractmethod
-    def measurement(self):
-        """
-        Return band information on the per-pixel metadata.
-        """
-
-
 class ObservedDaysSince(PerPixelMetadata):
     def __init__(self, since=datetime(1970, 1, 1), var_name='observed'):
         super(ObservedDaysSince, self).__init__(var_name)
@@ -684,372 +580,3 @@ class Medoid(Statistic):
         else:
             msg = 'Medoid<minimum_valid_observations={}>'
         return msg.format(self.minimum_valid_observations)
-
-
-class ExternalPlugin(Statistic):
-    """
-    Run externally defined plugin.
-    """
-
-    def __init__(self, impl, *args, **kwargs):
-        # Temporarily, add current path
-        sys.path.insert(0, os.getcwd())
-
-        LOG.debug('Looking for external plugin `%s` in %s', impl, sys.path)
-        impl_class = locate(impl)
-
-        # Remove the path that was added
-        sys.path = sys.path[1:]
-
-        if impl_class is None:
-            raise StatsProcessingError("Failed to load external plugin: '{}'".format(impl))
-        else:
-            LOG.debug('Found external plugin `%s`', impl)
-
-        self.impl = impl_class(*args, **kwargs)
-
-    def is_iterative(self) -> bool:
-        return self.impl.is_iterative()
-
-    def make_iterative_proc(self):
-        return self.impl.make_iterative_proc()
-
-    def measurements(self, input_measurements: Iterable[Measurement]) -> Iterable[Measurement]:
-        return self.impl.measurements(input_measurements)
-
-    def compute(self, data: xarray.Dataset) -> xarray.Dataset:
-        return self.impl.compute(data)
-
-    # caused trouble in unpickle stream
-    # def __getattr__(self, name):
-    #     # If attribute not on current object or on Statistic, try to find it on self.impl
-    #     return getattr(self.impl, name)
-
-
-class MaskMultiCounter(Statistic):
-    # pylint: disable=redefined-builtin
-    def __init__(self, vars, nodata_flags=None):
-        """
-        vars:
-           - name: <output_variable_name: String>
-             simple: <optional Bool, default: False>
-             flags:
-               field_name1: expected_value1
-               field_name2: expected_value2
-
-        # optional, define input nodata as a mask
-        # when all inputs match this, then output will be set to nodata
-        # this allows to distinguish 0 from nodata
-
-        nodata_flags:
-           contiguous: False
-
-        If variable is marked simple, then there is no distinction between 0 and nodata.
-        """
-        self._vars = [v.copy() for v in vars]
-        self._nodata_flags = nodata_flags
-        self._valid_pq_mask = None
-
-    def measurements(self, input_measurements):
-        nodata = -1
-        bit_defs = input_measurements[0].flags_definition
-
-        if self._nodata_flags is not None:
-            self._valid_pq_mask = mk_masker(*create_mask_value(bit_defs, **self._nodata_flags), invert=True)
-
-        for v in self._vars:
-            flags = v['flags']
-            v['_mask'] = create_mask_value(bit_defs, **flags)
-            v['mask'] = mk_masker(*v['_mask'])
-
-        return [Measurement(name=v['name'], dtype='int16', units='1', nodata=nodata)
-                for v in self._vars]
-
-    def is_iterative(self):
-        return True
-
-    def make_iterative_proc(self):
-        def _to_mask(ds):
-            da = first_var(ds)
-            return xarray.Dataset({v['name']: v['mask'](da) for v in self._vars},
-                                  attrs=ds.attrs)
-
-        # PQ -> BoolMasks: DataSet<Bool> -> Sum: DataSet<int16>
-        proc = compose_proc(_to_mask,
-                            proc=mk_incremental_sum(dtype='int16'))
-
-        if self._valid_pq_mask is None:
-            return proc
-
-        def invalid_data_mask(da):
-            mm = da.values
-            if mm.all():  # All pixels had at least one valid observation
-                return None
-            return np.logical_not(mm, out=mm)
-
-        # PQ -> ValidMask:DataArray<Bool> -> OR:DataArray<Bool> |> InvertMask:ndarray<Bool>|None
-        valid_proc = compose_proc(lambda ds: self._valid_pq_mask(first_var(ds)),
-                                  proc=mk_incremental_or(),
-                                  output_transform=invalid_data_mask)
-
-        _vars = {v['name']: v for v in self._vars}
-
-        def apply_mask(ds, mask, nodata=-1):
-            if mask is None:
-                return ds
-
-            for name, da in ds.data_vars.items():
-                simple = _vars[name].get('simple', False)
-                if not simple:
-                    da.values[mask] = nodata
-
-            return ds
-
-        # Counts      ----\
-        #                  +----------- Counts[ InvalidMask ] = nodata
-        # InvalidMask ----/
-
-        return broadcast_proc(proc, valid_proc, combine=apply_mask)
-
-    def compute(self, data):
-        proc = self.make_iterative_proc()
-
-        for i in range(data.time.shape[0]):
-            proc(data.isel(time=slice(i, i + 1)))
-
-        return proc()
-
-    def __repr__(self):
-        return 'MaskMultiCounter<{}>'.format(','.join([v['name'] for v in self._vars]))
-
-
-STATS = {
-    'simple': ReducingXarrayStatistic,
-    'percentile': Percentile,
-    'percentile_no_prov': PercentileNoProv,
-    'medoid': Medoid,
-    'medoid_no_prov': MedoidNoProv,
-    'medoid_simple': MedoidSimple,
-    'simple_normalised_difference': NormalisedDifferenceStats,
-    # 'ndvi_stats': NormalisedDifferenceStats(name='ndvi', band1='nir', band2='red', stats=['min', 'mean', 'max']),
-    # 'ndwi_stats': NormalisedDifferenceStats(name='ndwi', band1='green', band2='swir1', stats=['min', 'mean', 'max']),
-    # 'ndvi_daily': NormalisedDifferenceStats(name='ndvi', band1='nir', band2='red', stats=['squeeze']),
-    'none': NoneStat,
-    'wofs_summary': WofsStats,
-    'tcwbg_summary': TCWStats,
-    'masked_multi_count': MaskMultiCounter,
-    'external': ExternalPlugin,
-}
-
-
-# Dynamically look for and load statistics from other packages
-
-for entry_point in iter_entry_points(group='datacube.stats', name=None):
-    STATS[entry_point.name] = entry_point.load()
-
-try:
-    from hdmedians import nangeomedian
-    import warnings
-
-    def apply_geomedian(inarray, f, axis=3, eps=1e-3, **kwargs):
-        assert len(inarray.shape) == 4
-        assert axis == 3
-
-        xs, ys, bands, times = inarray.shape
-        output = np.ndarray((xs, ys, bands), dtype=inarray.dtype)
-        with warnings.catch_warnings():  # Don't print error about computing mean of empty slice
-            warnings.simplefilter("ignore", category=RuntimeWarning)
-            for ix in range(xs):
-                for iy in range(ys):
-                    try:
-                        output[ix, iy, :] = f(inarray[ix, iy, :, :], eps=eps, axis=1)
-                    except ValueError:
-                        output[ix, iy, :] = np.nan
-        return output
-
-    class GeoMedian(Statistic):
-        def __init__(self, eps=1e-3, maxiters=None):
-            super(GeoMedian, self).__init__()
-            self.eps = eps
-            self.maxiters = maxiters
-
-        def compute(self, data):
-            from_, to = self._vars_to_transpose(data)
-            # Assert data shape/dims
-            data = data.to_array(dim='variable').transpose(*from_).copy()
-
-            data = data.reduce(apply_geomedian, dim='time', keep_attrs=True, f=nangeomedian,
-                               eps=self.eps, maxiters=self.maxiters)
-
-            return data.transpose(*to).to_dataset(dim='variable')
-
-        @staticmethod
-        def _vars_to_transpose(data):
-            """
-            We need to be able to handle data given to use in either Geographic or Projected form.
-
-            The Data Cube provided xarrays will contain different dimensions, latitude/longitude or x/y, which means
-            the array reshaping takes different arguments.
-            """
-            is_proj = 'x' in data and 'y' in data
-            is_geo = 'longitude' in data and 'latitude' in data
-            if is_proj and is_geo:
-                raise StatsProcessingError(
-                    'Data to process contains both geographic and projected dimensions, unable to proceed')
-            elif not is_proj and not is_geo:
-                raise StatsProcessingError(
-                    'Data to process contains neither geographic nor projected dimensions, unable to proceed')
-            elif is_proj:
-                return ('x', 'y', 'variable', 'time'), ('variable', 'y', 'x')
-            else:
-                return ('longitude', 'latitude', 'variable', 'time'), ('variable', 'latitude', 'longitude')
-
-    STATS['geomedian'] = GeoMedian
-
-except ImportError:
-    pass
-
-try:
-    from pcm import gmpcm
-
-    class NewGeomedianStatistic(Statistic):
-        def __init__(self, eps=1e-3, num_threads=None):
-            super(NewGeomedianStatistic, self).__init__()
-            self.eps = eps
-            self.num_threads = num_threads
-
-        def compute(self, data):
-            """
-            :param xarray.Dataset data:
-            :return: xarray.Dataset
-            """
-            # We need to reshape our data into Y, X, Band, Time
-
-            squashed_together_dimensions, normal_datacube_dimensions = self._vars_to_transpose(data)
-
-            squashed = data.to_array(dim='variable').transpose(*squashed_together_dimensions)
-            assert squashed.dims == squashed_together_dimensions
-
-            # Grab a copy of the coordinates we need for creating the output DataArray
-            output_coords = copy(squashed.coords)
-            if 'time' in output_coords:
-                del output_coords['time']
-            if 'source' in output_coords:
-                del output_coords['source']
-
-            # Call Dale's function here
-            squashed = gmpcm(squashed.data, num_threads=self.num_threads)
-            all_zeros = (squashed == 0.).all(axis=-1)
-            squashed[all_zeros] = np.nan
-
-            # Jam the raw numpy array back into a pleasantly labelled DataArray
-            output_dims = squashed_together_dimensions[:-1]
-            as_datarray = xarray.DataArray(squashed, dims=output_dims, coords=output_coords)
-
-            return as_datarray.transpose(*normal_datacube_dimensions).to_dataset(dim='variable')
-
-        @staticmethod
-        def _vars_to_transpose(data):
-            """
-            We need to be able to handle data given to use in either Geographic or Projected form.
-
-            The Data Cube provided xarrays will contain different dimensions, latitude/longitude or x/y, which means
-            the array reshaping takes different arguments.
-
-            The dimension ordering returned by this function is specific to the Geometric Median PCM functions
-            included from the `pcm` module.
-
-            :return: pcm input array dimension order, datacube dimension ordering
-            """
-            is_projected = 'x' in data.dims and 'y' in data.dims
-            is_geographic = 'longitude' in data.dims and 'latitude' in data.dims
-
-            if is_projected and is_geographic:
-                raise StatsProcessingError('Data to process contains BOTH geographic and projected dimensions, '
-                                           'unable to proceed')
-            elif not is_projected and not is_geographic:
-                raise StatsProcessingError('Data to process contains NEITHER geographic nor projected dimensions, '
-                                           'unable to proceed')
-            elif is_projected:
-                return ('y', 'x', 'variable', 'time'), ('variable', 'y', 'x')
-            else:
-                return ('latitude', 'longitude', 'variable', 'time'), ('variable', 'latitude', 'longitude')
-
-    STATS['new_geomedian'] = NewGeomedianStatistic
-
-except ImportError:
-    pass
-
-
-try:
-    import pcm
-
-    class SpectralMAD(Statistic):
-        def __init__(self, eps=1e-3, num_threads=None):
-            super(SpectralMAD, self).__init__()
-            self.eps = eps
-            self.num_threads = num_threads
-
-        def compute(self, data):
-            """
-            :param xarray.Dataset data:
-            :return: xarray.Dataset
-            """
-
-            # We need to reshape our data into Y, X, Band, Time
-            squashed_together_dimensions, output_dimensions = self._vars_to_transpose(data)
-
-            squashed = data.to_array(dim='variable').transpose(*squashed_together_dimensions)
-            assert squashed.dims == squashed_together_dimensions
-
-            # Grab a copy of the coordinates we need for creating the output DataArray
-            output_coords = copy(squashed.coords)
-            if 'variable' in output_coords:
-                del output_coords['variable']
-            if 'time' in output_coords:
-                del output_coords['time']
-            if 'source' in output_coords:
-                del output_coords['source']
-
-            # Call Dale's geometric median & spectral mad functions here
-            gm = pcm.gmpcm(squashed.data, num_threads=self.num_threads)
-            squashed = pcm.smad(squashed.data, gm, num_threads=self.num_threads)
-
-            # Jam the raw numpy array back into a pleasantly labelled DataArray
-            as_datarray = xarray.DataArray(squashed, dims=output_dimensions, coords=output_coords)
-
-            return as_datarray.transpose(*output_dimensions).to_dataset(name='smad')
-
-        def measurements(self, input_measurements):
-            return [Measurement(name='smad', dtype='float32', nodata=np.nan, units='1')]
-
-        @staticmethod
-        def _vars_to_transpose(data):
-            """
-            We need to be able to handle data given to use in either Geographic or Projected form.
-
-            The Data Cube provided xarrays will contain different dimensions, latitude/longitude or x/y, which means
-            the array reshaping takes different arguments.
-
-            The dimension ordering returned by this function is specific to the spectral median absolute deviation
-            function included from the `pcm` module.
-
-            :return: pcm input array dimension order, datacube dimension ordering
-            """
-            is_projected = 'x' in data.dims and 'y' in data.dims
-            is_geographic = 'longitude' in data.dims and 'latitude' in data.dims
-
-            if is_projected and is_geographic:
-                raise StatsProcessingError('Data to process contains BOTH geographic and projected dimensions, '
-                                           'unable to proceed')
-            elif not is_projected and not is_geographic:
-                raise StatsProcessingError('Data to process contains NEITHER geographic nor projected dimensions, '
-                                           'unable to proceed')
-            elif is_projected:
-                return ('y', 'x', 'variable', 'time'), ('y', 'x')
-            else:
-                return ('latitude', 'longitude', 'variable', 'time'), ('variable', 'latitude', 'longitude')
-
-    STATS['spectral_mad'] = SpectralMAD
-except ImportError:
-    pass
