+++ setup.py
@@ -27,9 +27,6 @@ setup(
         'console_scripts': [
             'datacube-stats = datacube_stats.main:main',
             'datacube-tile-check = datacube_stats.cli.tile_check:main',
-        ],
-        'datacube.stats': [
-            'wofs-summary = datacube_stats.statistics:WofsStats'
         ]
     },
 )
